PY3:  Many Python 3 dictionary fixes via the 2to3 Python script.
The command used was:
2to3 -j 4 -w -f dict .

The keys() dictionary method in Python 3 now returns an iterator object,
so this needs to be passed through the list() builtin function.  The
iteritems() dictionary method no longer exists, so items() needs to be
used instead (sometimes passed through list() when lists rather than
tuples are required).